feat: GET /users 구현

+++ src/main/java/cart/dto/UserResponse.java
@@ -3,10 +3,10 @@ package cart.dto;
 public class UserResponse {
 
     private final String email;
-    private final Long point;
+    private final int point;
     private final double earnRate;
 
-    public UserResponse(final String email, final Long point, final double earnRate) {
+    public UserResponse(final String email, final int point, final double earnRate) {
         this.email = email;
         this.point = point;
         this.earnRate = earnRate;
@@ -16,7 +16,7 @@ public class UserResponse {
         return email;
     }
 
-    public Long getPoint() {
+    public int getPoint() {
         return point;
     }
